DikuMUD alfa release
This is the second and final release, from July 1991. The contents are
from an archive named "dikumud.tar.gz" and contains files timestamped
Nov 9 1991. Some of this will be the Linux port changes reverting back
to the original gamma release.

+++ reception.c
@@ -203,8 +203,12 @@ void load_char_objs(struct char_data *ch)
 	if (found) {
 
 		obj_store_to_char(ch, &st);
-		
-		timegold = (int) ((st.total_cost*(time(0) - st.last_update)) / (SECS_PER_REAL_DAY));
+
+		/* To avoid overflow of int */
+
+		timegold =
+       (unsigned) (((float) st.total_cost*(float) (time(0) - st.last_update))
+                  / (float) (SECS_PER_REAL_DAY));
 
 		GET_GOLD(ch) -= timegold;
